pyproject.toml
[project]
name = "trajectory-prediction-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Lucas Cavalcante dos Santos",email = "cavalcanteprofissional@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "catboost (>=1.0.0)",
    "downloader (>=0.98,<0.99)",
    "kaggle (>=1.5.0)",
    "lightgbm (>=3.3.0)",
    "matplotlib (>=3.7.0)",
    "numpy (>=1.24.0)",
    "pandas (>=2.0.0)",
    "python-dotenv (>=1.0.0)",
    "scikit-learn (>=1.3.0)",
    "seaborn (>=0.12.0)",
    "xgboost (>=1.7.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
